tracee-ebpf: restructure and split files
