+++ tracee-ebpf/tracee/net_events.go
@@ -1,0 +1,144 @@
+package tracee
+
+import (
+	"bytes"
+	"encoding/binary"
+	"fmt"
+	"time"
+
+	"github.com/google/gopacket"
+	"inet.af/netaddr"
+)
+
+func (t *Tracee) processNetEvents() {
+	// Todo: split pcap files by context (tid + comm)
+	// Todo: add stats for network packets (in epilog)
+	for {
+		select {
+		case in := <-t.netChannel:
+			// Sanity check - timestamp, event id, host tid and comm must exist in all net events
+			if len(in) < 32 {
+				continue
+			}
+
+			timeStamp := binary.LittleEndian.Uint64(in[0:8])
+			netEventId := binary.LittleEndian.Uint32(in[8:12])
+			hostTid := binary.LittleEndian.Uint32(in[12:16])
+			comm := string(bytes.TrimRight(in[16:32], "\x00"))
+			dataBuff := bytes.NewBuffer(in[32:])
+
+			// timeStamp is nanoseconds since system boot time
+			timeStampObj := time.Unix(0, int64(timeStamp+t.bootTime))
+
+			if netEventId == NetPacket {
+				var pktLen uint32
+				err := binary.Read(dataBuff, binary.LittleEndian, &pktLen)
+				if err != nil {
+					t.handleError(err)
+					continue
+				}
+
+				if t.config.Debug {
+					var pktMeta struct {
+						SrcIP    [16]byte
+						DestIP   [16]byte
+						SrcPort  uint16
+						DestPort uint16
+						Protocol uint8
+						_        [7]byte //padding
+					}
+					err = binary.Read(dataBuff, binary.LittleEndian, &pktMeta)
+					if err != nil {
+						t.handleError(err)
+						continue
+					}
+
+					fmt.Printf("%v  %-16s  %-7d  debug_net/packet               Len: %d, SrcIP: %v, SrcPort: %d, DestIP: %v, DestPort: %d, Protocol: %d\n",
+						timeStampObj,
+						comm,
+						hostTid,
+						pktLen,
+						netaddr.IPFrom16(pktMeta.SrcIP),
+						pktMeta.SrcPort,
+						netaddr.IPFrom16(pktMeta.DestIP),
+						pktMeta.DestPort,
+						pktMeta.Protocol)
+				}
+
+				info := gopacket.CaptureInfo{
+					Timestamp:      timeStampObj,
+					CaptureLength:  int(pktLen),
+					Length:         int(pktLen),
+					InterfaceIndex: 0, // todo: accept array of interfaces?
+				}
+
+				err = t.pcapWriter.WritePacket(info, dataBuff.Bytes()[:pktLen])
+				if err != nil {
+					t.handleError(err)
+					continue
+				}
+
+				// todo: maybe we should not flush every packet?
+				err = t.pcapWriter.Flush()
+				if err != nil {
+					t.handleError(err)
+					continue
+				}
+			} else if t.config.Debug {
+				var pkt struct {
+					LocalIP     [16]byte
+					RemoteIP    [16]byte
+					LocalPort   uint16
+					RemotePort  uint16
+					Protocol    uint8
+					_           [3]byte //padding
+					TcpOldState uint32
+					TcpNewState uint32
+					_           [4]byte //padding
+					SockPtr     uint64
+				}
+				err := binary.Read(dataBuff, binary.LittleEndian, &pkt)
+				if err != nil {
+					t.handleError(err)
+					continue
+				}
+
+				switch netEventId {
+				case DebugNetSecurityBind:
+					fmt.Printf("%v  %-16s  %-7d  debug_net/security_socket_bind LocalIP: %v, LocalPort: %d, Protocol: %d\n",
+						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
+				case DebugNetUdpSendmsg:
+					fmt.Printf("%v  %-16s  %-7d  debug_net/udp_sendmsg          LocalIP: %v, LocalPort: %d, Protocol: %d\n",
+						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
+				case DebugNetUdpDisconnect:
+					fmt.Printf("%v  %-16s  %-7d  debug_net/__udp_disconnect     LocalIP: %v, LocalPort: %d, Protocol: %d\n",
+						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
+				case DebugNetUdpDestroySock:
+					fmt.Printf("%v  %-16s  %-7d  debug_net/udp_destroy_sock     LocalIP: %v, LocalPort: %d, Protocol: %d\n",
+						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
+				case DebugNetUdpV6DestroySock:
+					fmt.Printf("%v  %-16s  %-7d  debug_net/udpv6_destroy_sock   LocalIP: %v, LocalPort: %d, Protocol: %d\n",
+						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
+				case DebugNetInetSockSetState:
+					fmt.Printf("%v  %-16s  %-7d  debug_net/inet_sock_set_state  LocalIP: %v, LocalPort: %d, RemoteIP: %v, RemotePort: %d, Protocol: %d, OldState: %d, NewState: %d, SockPtr: 0x%x\n",
+						timeStampObj,
+						comm,
+						hostTid,
+						netaddr.IPFrom16(pkt.LocalIP),
+						pkt.LocalPort,
+						netaddr.IPFrom16(pkt.RemoteIP),
+						pkt.RemotePort,
+						pkt.Protocol,
+						pkt.TcpOldState,
+						pkt.TcpNewState,
+						pkt.SockPtr)
+				case DebugNetTcpConnect:
+					fmt.Printf("%v  %-16s  %-7d  debug_net/tcp_connect     LocalIP: %v, LocalPort: %d, Protocol: %d\n",
+						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
+				}
+			}
+		case lost := <-t.lostNetChannel:
+			t.stats.lostNtCounter.Increment(int(lost))
+		}
+	}
+}
