+++ tracee-ebpf/tracee/tracee.go
@@ -1,9 +1,7 @@
 package tracee
 
 import (
-	"bytes"
 	"crypto/sha256"
-	"encoding/binary"
 	"encoding/hex"
 	"encoding/json"
 	"errors"
@@ -25,11 +23,9 @@ import (
 	bpf "github.com/aquasecurity/libbpfgo"
 	"github.com/aquasecurity/libbpfgo/helpers"
 	"github.com/aquasecurity/tracee/tracee-ebpf/external"
-	"github.com/google/gopacket"
 	"github.com/google/gopacket/layers"
 	"github.com/google/gopacket/pcapgo"
 	"golang.org/x/sys/unix"
-	"inet.af/netaddr"
 )
 
 // Config is a struct containing user defined configuration of tracee
@@ -449,58 +445,6 @@ func New(cfg Config) (*Tracee, error) {
 	t.StackAddressesMap = StackAddressesMap
 
 	return t, nil
-}
-
-type bucketsCache struct {
-	buckets     map[uint32][]uint32
-	bucketLimit int
-	Null        uint32
-}
-
-func (c *bucketsCache) Init(bucketLimit int) {
-	c.bucketLimit = bucketLimit
-	c.buckets = make(map[uint32][]uint32)
-	c.Null = 0
-}
-
-func (c *bucketsCache) GetBucket(key uint32) []uint32 {
-	return c.buckets[key]
-}
-
-func (c *bucketsCache) GetBucketItem(key uint32, index int) uint32 {
-	b, exists := c.buckets[key]
-	if !exists {
-		return c.Null
-	}
-	if index >= len(b) {
-		return c.Null
-	}
-	return b[index]
-}
-
-func (c *bucketsCache) AddBucketItem(key uint32, value uint32) {
-	c.addBucketItem(key, value, false)
-}
-
-func (c *bucketsCache) ForceAddBucketItem(key uint32, value uint32) {
-	c.addBucketItem(key, value, true)
-}
-
-func (c *bucketsCache) addBucketItem(key uint32, value uint32, force bool) {
-	b, exists := c.buckets[key]
-	if !exists {
-		c.buckets[key] = make([]uint32, 0, c.bucketLimit)
-		b = c.buckets[key]
-	}
-	if len(b) >= c.bucketLimit {
-		if force {
-			b[0] = value
-		} else {
-			return
-		}
-	} else {
-		c.buckets[key] = append(b, value)
-	}
 }
 
 // UnameRelease gets the version string of the current running kernel
@@ -1284,196 +1228,9 @@ func boolToUInt32(b bool) uint32 {
 	return uint32(0)
 }
 
-// CopyFileByPath copies a file from src to dst
-func CopyFileByPath(src, dst string) error {
-	sourceFileStat, err := os.Stat(src)
-	if err != nil {
-		return err
-	}
-	if !sourceFileStat.Mode().IsRegular() {
-		return fmt.Errorf("%s is not a regular file", src)
-	}
-	source, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer source.Close()
-	destination, err := os.Create(dst)
-	if err != nil {
-		return err
-	}
-	defer destination.Close()
-	_, err = io.Copy(destination, source)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
 func (t *Tracee) handleError(err error) {
 	t.stats.errorCounter.Increment()
 	t.printer.Error(err)
-}
-
-// shouldProcessEvent decides whether or not to drop an event before further processing it
-func (t *Tracee) shouldProcessEvent(e RawEvent) bool {
-	if t.config.Filter.RetFilter.Enabled {
-		if filter, ok := t.config.Filter.RetFilter.Filters[e.Ctx.EventID]; ok {
-			retVal := e.Ctx.Retval
-			match := false
-			for _, f := range filter.Equal {
-				if retVal == f {
-					match = true
-					break
-				}
-			}
-			if !match && len(filter.Equal) > 0 {
-				return false
-			}
-			for _, f := range filter.NotEqual {
-				if retVal == f {
-					return false
-				}
-			}
-			if (filter.Greater != GreaterNotSetInt) && retVal <= filter.Greater {
-				return false
-			}
-			if (filter.Less != LessNotSetInt) && retVal >= filter.Less {
-				return false
-			}
-		}
-	}
-
-	if t.config.Filter.ArgFilter.Enabled {
-		for _, filter := range t.config.Filter.ArgFilter.Filters[e.Ctx.EventID] {
-			argVal, ok := e.RawArgs[filter.argTag]
-			if !ok {
-				continue
-			}
-			// TODO: use type assertion instead of string convertion
-			argValStr := fmt.Sprint(argVal)
-			match := false
-			for _, f := range filter.Equal {
-				if argValStr == f || (f[len(f)-1] == '*' && strings.HasPrefix(argValStr, f[0:len(f)-1])) {
-					match = true
-					break
-				}
-			}
-			if !match && len(filter.Equal) > 0 {
-				return false
-			}
-			for _, f := range filter.NotEqual {
-				if argValStr == f || (f[len(f)-1] == '*' && strings.HasPrefix(argValStr, f[0:len(f)-1])) {
-					return false
-				}
-			}
-		}
-	}
-
-	return true
-}
-
-func (t *Tracee) processEvent(ctx *context, args map[argTag]interface{}) error {
-	switch ctx.EventID {
-
-	//capture written files
-	case VfsWriteEventID, VfsWritevEventID:
-		if t.config.Capture.FileWrite {
-			filePath, ok := args[t.EncParamName[ctx.EventID%2]["pathname"]].(string)
-			if !ok {
-				return fmt.Errorf("error parsing vfs_write args")
-			}
-			// path should be absolute, except for e.g memfd_create files
-			if filePath == "" || filePath[0] != '/' {
-				return nil
-			}
-			dev, ok := args[t.EncParamName[ctx.EventID%2]["dev"]].(uint32)
-			if !ok {
-				return fmt.Errorf("error parsing vfs_write args")
-			}
-			inode, ok := args[t.EncParamName[ctx.EventID%2]["inode"]].(uint64)
-			if !ok {
-				return fmt.Errorf("error parsing vfs_write args")
-			}
-
-			// stop processing if write was already indexed
-			fileName := fmt.Sprintf("%d/write.dev-%d.inode-%d", ctx.MntID, dev, inode)
-			indexName, ok := t.writtenFiles[fileName]
-			if ok && indexName == filePath {
-				return nil
-			}
-
-			// index written file by original filepath
-			t.writtenFiles[fileName] = filePath
-		}
-
-	case SecurityBprmCheckEventID:
-
-		//cache this pid by it's mnt ns
-		if ctx.Pid == 1 {
-			t.pidsInMntns.ForceAddBucketItem(ctx.MntID, ctx.HostPid)
-		} else {
-			t.pidsInMntns.AddBucketItem(ctx.MntID, ctx.HostPid)
-		}
-
-		//capture executed files
-		if t.config.Capture.Exec {
-			filePath, ok := args[t.EncParamName[ctx.EventID%2]["pathname"]].(string)
-			if !ok {
-				return fmt.Errorf("error parsing security_bprm_check args")
-			}
-			// path should be absolute, except for e.g memfd_create files
-			if filePath == "" || filePath[0] != '/' {
-				return nil
-			}
-
-			destinationDirPath := filepath.Join(t.config.Capture.OutputPath, strconv.Itoa(int(ctx.MntID)))
-			if err := os.MkdirAll(destinationDirPath, 0755); err != nil {
-				return err
-			}
-			destinationFilePath := filepath.Join(destinationDirPath, fmt.Sprintf("exec.%d.%s", ctx.Ts, filepath.Base(filePath)))
-
-			var err error
-			// try to access the root fs via another process in the same mount namespace (since the current process might have already died)
-			pids := t.pidsInMntns.GetBucket(ctx.MntID)
-			for _, pid := range pids { // will break on success
-				err = nil
-				sourceFilePath := fmt.Sprintf("/proc/%s/root%s", strconv.Itoa(int(pid)), filePath)
-				var sourceFileStat os.FileInfo
-				sourceFileStat, err = os.Stat(sourceFilePath)
-				if err != nil {
-					//TODO: remove dead pid from cache
-					continue
-				}
-
-				sourceFileCtime := sourceFileStat.Sys().(*syscall.Stat_t).Ctim.Nano()
-				capturedFileID := fmt.Sprintf("%d:%s", ctx.MntID, sourceFilePath)
-
-				// create an in-memory profile
-				if t.config.Capture.Profile {
-					t.updateProfile(fmt.Sprintf("%s:%d", filepath.Join(destinationDirPath, fmt.Sprintf("exec.%s", filepath.Base(filePath))), sourceFileCtime), ctx.Ts)
-				}
-
-				//don't capture same file twice unless it was modified
-				lastCtime, ok := t.capturedFiles[capturedFileID]
-				if ok && lastCtime == sourceFileCtime {
-					return nil
-				}
-
-				//capture
-				err = CopyFileByPath(sourceFilePath, destinationFilePath)
-				if err != nil {
-					return err
-				}
-				//mark this file as captured
-				t.capturedFiles[capturedFileID] = sourceFileCtime
-				break
-			}
-			return err
-		}
-	}
-
-	return nil
 }
 
 func getFileHash(fileName string) string {
@@ -1485,18 +1242,6 @@ func getFileHash(fileName string) string {
 		return hex.EncodeToString(h.Sum(nil))
 	}
 	return ""
-}
-
-func (t *Tracee) updateProfile(sourceFilePath string, executionTs uint64) {
-	if pf, ok := t.profiledFiles[sourceFilePath]; !ok {
-		t.profiledFiles[sourceFilePath] = profilerInfo{
-			Times:            1,
-			FirstExecutionTs: executionTs,
-		}
-	} else {
-		pf.Times = pf.Times + 1              // bump execution count
-		t.profiledFiles[sourceFilePath] = pf // update
-	}
 }
 
 func (t *Tracee) updateFileSHA() {
@@ -1646,550 +1391,4 @@ func (t *Tracee) prepareArgsForPrint(ctx *context, args map[argTag]interface{}) 
 	}
 
 	return nil
-}
-
-// context struct contains common metadata that is collected for all types of events
-// it is used to unmarshal binary data and therefore should match (bit by bit) to the `context_t` struct in the ebpf code.
-// NOTE: Integers want to be aligned in memory, so if changing the format of this struct
-// keep the 1-byte 'Argnum' as the final parameter before the padding (if padding is needed).
-type context struct {
-	Ts       uint64
-	Pid      uint32
-	Tid      uint32
-	Ppid     uint32
-	HostPid  uint32
-	HostTid  uint32
-	HostPpid uint32
-	Uid      uint32
-	MntID    uint32
-	PidID    uint32
-	Comm     [16]byte
-	UtsName  [16]byte
-	ContID   [16]byte
-	EventID  int32
-	Retval   int64
-	StackID  uint32
-	Argnum   uint8
-	_        [3]byte //padding
-}
-
-func (t *Tracee) processLostEvents() {
-	for {
-		lost := <-t.lostEvChannel
-		t.stats.lostEvCounter.Increment(int(lost))
-	}
-}
-
-func (t *Tracee) processFileWrites() {
-	type chunkMeta struct {
-		BinType  binType
-		MntID    uint32
-		Metadata [20]byte
-		Size     int32
-		Off      uint64
-	}
-
-	type vfsWriteMeta struct {
-		DevID uint32
-		Inode uint64
-		Mode  uint32
-		Pid   uint32
-	}
-
-	type mprotectWriteMeta struct {
-		Ts uint64
-	}
-
-	const (
-		S_IFMT uint32 = 0170000 // bit mask for the file type bit field
-
-		S_IFSOCK uint32 = 0140000 // socket
-		S_IFLNK  uint32 = 0120000 // symbolic link
-		S_IFREG  uint32 = 0100000 // regular file
-		S_IFBLK  uint32 = 0060000 // block device
-		S_IFDIR  uint32 = 0040000 // directory
-		S_IFCHR  uint32 = 0020000 // character device
-		S_IFIFO  uint32 = 0010000 // FIFO
-	)
-
-	for {
-		select {
-		case dataRaw := <-t.fileWrChannel:
-			if len(dataRaw) == 0 {
-				continue
-			}
-			dataBuff := bytes.NewBuffer(dataRaw)
-			var meta chunkMeta
-			appendFile := false
-			err := binary.Read(dataBuff, binary.LittleEndian, &meta)
-			if err != nil {
-				t.handleError(err)
-				continue
-			}
-
-			if meta.Size <= 0 {
-				t.handleError(fmt.Errorf("error in file writer: invalid chunk size: %d", meta.Size))
-				continue
-			}
-			if dataBuff.Len() < int(meta.Size) {
-				t.handleError(fmt.Errorf("error in file writer: chunk too large: %d", meta.Size))
-				continue
-			}
-
-			pathname := path.Join(t.config.Capture.OutputPath, strconv.Itoa(int(meta.MntID)))
-			if err := os.MkdirAll(pathname, 0755); err != nil {
-				t.handleError(err)
-				continue
-			}
-			filename := ""
-			metaBuff := bytes.NewBuffer(meta.Metadata[:])
-			if meta.BinType == sendVfsWrite {
-				var vfsMeta vfsWriteMeta
-				err = binary.Read(metaBuff, binary.LittleEndian, &vfsMeta)
-				if err != nil {
-					t.handleError(err)
-					continue
-				}
-				if vfsMeta.Mode&S_IFSOCK == S_IFSOCK || vfsMeta.Mode&S_IFCHR == S_IFCHR || vfsMeta.Mode&S_IFIFO == S_IFIFO {
-					appendFile = true
-				}
-				if vfsMeta.Pid == 0 {
-					filename = fmt.Sprintf("write.dev-%d.inode-%d", vfsMeta.DevID, vfsMeta.Inode)
-				} else {
-					filename = fmt.Sprintf("write.dev-%d.inode-%d.pid-%d", vfsMeta.DevID, vfsMeta.Inode, vfsMeta.Pid)
-				}
-			} else if meta.BinType == sendMprotect {
-				var mprotectMeta mprotectWriteMeta
-				err = binary.Read(metaBuff, binary.LittleEndian, &mprotectMeta)
-				if err != nil {
-					t.handleError(err)
-					continue
-				}
-				// note: size of buffer will determine maximum extracted file size! (as writes from kernel are immediate)
-				filename = fmt.Sprintf("bin.%d", mprotectMeta.Ts)
-			} else {
-				t.handleError(fmt.Errorf("error in file writer: unknown binary type: %d", meta.BinType))
-				continue
-			}
-
-			fullname := path.Join(pathname, filename)
-
-			f, err := os.OpenFile(fullname, os.O_CREATE|os.O_WRONLY, 0640)
-			if err != nil {
-				t.handleError(err)
-				continue
-			}
-			if appendFile {
-				if _, err := f.Seek(0, io.SeekEnd); err != nil {
-					f.Close()
-					t.handleError(err)
-					continue
-				}
-			} else {
-				if _, err := f.Seek(int64(meta.Off), io.SeekStart); err != nil {
-					f.Close()
-					t.handleError(err)
-					continue
-				}
-			}
-
-			dataBytes, err := readByteSliceFromBuff(dataBuff, int(meta.Size))
-			if err != nil {
-				f.Close()
-				t.handleError(err)
-				continue
-			}
-			if _, err := f.Write(dataBytes); err != nil {
-				f.Close()
-				t.handleError(err)
-				continue
-			}
-			if err := f.Close(); err != nil {
-				t.handleError(err)
-				continue
-			}
-		case lost := <-t.lostWrChannel:
-			t.stats.lostWrCounter.Increment(int(lost))
-		}
-	}
-}
-
-func (t *Tracee) processNetEvents() {
-	// Todo: split pcap files by context (tid + comm)
-	// Todo: add stats for network packets (in epilog)
-	for {
-		select {
-		case in := <-t.netChannel:
-			// Sanity check - timestamp, event id, host tid and comm must exist in all net events
-			if len(in) < 32 {
-				continue
-			}
-
-			timeStamp := binary.LittleEndian.Uint64(in[0:8])
-			netEventId := binary.LittleEndian.Uint32(in[8:12])
-			hostTid := binary.LittleEndian.Uint32(in[12:16])
-			comm := string(bytes.TrimRight(in[16:32], "\x00"))
-			dataBuff := bytes.NewBuffer(in[32:])
-
-			// timeStamp is nanoseconds since system boot time
-			timeStampObj := time.Unix(0, int64(timeStamp+t.bootTime))
-
-			if netEventId == NetPacket {
-				var pktLen uint32
-				err := binary.Read(dataBuff, binary.LittleEndian, &pktLen)
-				if err != nil {
-					t.handleError(err)
-					continue
-				}
-
-				if t.config.Debug {
-					var pktMeta struct {
-						SrcIP    [16]byte
-						DestIP   [16]byte
-						SrcPort  uint16
-						DestPort uint16
-						Protocol uint8
-						_        [7]byte //padding
-					}
-					err = binary.Read(dataBuff, binary.LittleEndian, &pktMeta)
-					if err != nil {
-						t.handleError(err)
-						continue
-					}
-
-					fmt.Printf("%v  %-16s  %-7d  debug_net/packet               Len: %d, SrcIP: %v, SrcPort: %d, DestIP: %v, DestPort: %d, Protocol: %d\n",
-						timeStampObj,
-						comm,
-						hostTid,
-						pktLen,
-						netaddr.IPFrom16(pktMeta.SrcIP),
-						pktMeta.SrcPort,
-						netaddr.IPFrom16(pktMeta.DestIP),
-						pktMeta.DestPort,
-						pktMeta.Protocol)
-				}
-
-				info := gopacket.CaptureInfo{
-					Timestamp:      timeStampObj,
-					CaptureLength:  int(pktLen),
-					Length:         int(pktLen),
-					InterfaceIndex: 0, // todo: accept array of interfaces?
-				}
-
-				err = t.pcapWriter.WritePacket(info, dataBuff.Bytes()[:pktLen])
-				if err != nil {
-					t.handleError(err)
-					continue
-				}
-
-				// todo: maybe we should not flush every packet?
-				err = t.pcapWriter.Flush()
-				if err != nil {
-					t.handleError(err)
-					continue
-				}
-			} else if t.config.Debug {
-				var pkt struct {
-					LocalIP     [16]byte
-					RemoteIP    [16]byte
-					LocalPort   uint16
-					RemotePort  uint16
-					Protocol    uint8
-					_           [3]byte //padding
-					TcpOldState uint32
-					TcpNewState uint32
-					_           [4]byte //padding
-					SockPtr     uint64
-				}
-				err := binary.Read(dataBuff, binary.LittleEndian, &pkt)
-				if err != nil {
-					t.handleError(err)
-					continue
-				}
-
-				switch netEventId {
-				case DebugNetSecurityBind:
-					fmt.Printf("%v  %-16s  %-7d  debug_net/security_socket_bind LocalIP: %v, LocalPort: %d, Protocol: %d\n",
-						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
-				case DebugNetUdpSendmsg:
-					fmt.Printf("%v  %-16s  %-7d  debug_net/udp_sendmsg          LocalIP: %v, LocalPort: %d, Protocol: %d\n",
-						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
-				case DebugNetUdpDisconnect:
-					fmt.Printf("%v  %-16s  %-7d  debug_net/__udp_disconnect     LocalIP: %v, LocalPort: %d, Protocol: %d\n",
-						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
-				case DebugNetUdpDestroySock:
-					fmt.Printf("%v  %-16s  %-7d  debug_net/udp_destroy_sock     LocalIP: %v, LocalPort: %d, Protocol: %d\n",
-						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
-				case DebugNetUdpV6DestroySock:
-					fmt.Printf("%v  %-16s  %-7d  debug_net/udpv6_destroy_sock   LocalIP: %v, LocalPort: %d, Protocol: %d\n",
-						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
-				case DebugNetInetSockSetState:
-					fmt.Printf("%v  %-16s  %-7d  debug_net/inet_sock_set_state  LocalIP: %v, LocalPort: %d, RemoteIP: %v, RemotePort: %d, Protocol: %d, OldState: %d, NewState: %d, SockPtr: 0x%x\n",
-						timeStampObj,
-						comm,
-						hostTid,
-						netaddr.IPFrom16(pkt.LocalIP),
-						pkt.LocalPort,
-						netaddr.IPFrom16(pkt.RemoteIP),
-						pkt.RemotePort,
-						pkt.Protocol,
-						pkt.TcpOldState,
-						pkt.TcpNewState,
-						pkt.SockPtr)
-				case DebugNetTcpConnect:
-					fmt.Printf("%v  %-16s  %-7d  debug_net/tcp_connect     LocalIP: %v, LocalPort: %d, Protocol: %d\n",
-						timeStampObj, comm, hostTid, netaddr.IPFrom16(pkt.LocalIP), pkt.LocalPort, pkt.Protocol)
-				}
-			}
-		case lost := <-t.lostNetChannel:
-			t.stats.lostNtCounter.Increment(int(lost))
-		}
-	}
-}
-
-func readStringFromBuff(buff io.Reader) (string, error) {
-	var err error
-	var size uint32
-	err = binary.Read(buff, binary.LittleEndian, &size)
-	if err != nil {
-		return "", fmt.Errorf("error reading string size: %v", err)
-	}
-	if size > 4096 {
-		return "", fmt.Errorf("string size too big: %d", size)
-	}
-	res, err := readByteSliceFromBuff(buff, int(size-1)) //last byte is string terminating null
-	defer func() {
-		var dummy int8
-		binary.Read(buff, binary.LittleEndian, &dummy) //discard last byte which is string terminating null
-	}()
-	if err != nil {
-		return "", fmt.Errorf("error reading string arg: %v", err)
-	}
-	return string(res), nil
-}
-
-// readStringVarFromBuff reads a null-terminated string from `buff`
-// max length can be passed as `max` to optimize memory allocation, otherwise pass 0
-func readStringVarFromBuff(buff io.Reader, max int) (string, error) {
-	var err error
-	var char int8
-	res := make([]byte, max)
-	err = binary.Read(buff, binary.LittleEndian, &char)
-	if err != nil {
-		return "", fmt.Errorf("error reading null terminated string: %v", err)
-	}
-	for count := 1; char != 0 && count < max; count++ {
-		res = append(res, byte(char))
-		err = binary.Read(buff, binary.LittleEndian, &char)
-		if err != nil {
-			return "", fmt.Errorf("error reading null terminated string: %v", err)
-		}
-	}
-	res = bytes.TrimLeft(res[:], "\000")
-	return string(res), nil
-}
-
-func readByteSliceFromBuff(buff io.Reader, len int) ([]byte, error) {
-	var err error
-	res := make([]byte, len)
-	err = binary.Read(buff, binary.LittleEndian, &res)
-	if err != nil {
-		return nil, fmt.Errorf("error reading byte array: %v", err)
-	}
-	return res, nil
-}
-
-func readSockaddrFromBuff(buff io.Reader) (map[string]string, error) {
-	res := make(map[string]string, 3)
-	var family int16
-	err := binary.Read(buff, binary.LittleEndian, &family)
-	if err != nil {
-		return nil, err
-	}
-	res["sa_family"] = helpers.ParseSocketDomain(uint32(family))
-	switch family {
-	case 1: // AF_UNIX
-		/*
-			http://man7.org/linux/man-pages/man7/unix.7.html
-			struct sockaddr_un {
-					sa_family_t sun_family;     // AF_UNIX
-					char        sun_path[108];  // Pathname
-			};
-		*/
-		var sunPathBuf [108]byte
-		err := binary.Read(buff, binary.LittleEndian, &sunPathBuf)
-		if err != nil {
-			return nil, fmt.Errorf("error parsing sockaddr_un: %v", err)
-		}
-		trimmedPath := bytes.TrimLeft(sunPathBuf[:], "\000")
-		sunPath := ""
-		if len(trimmedPath) != 0 {
-			sunPath, err = readStringVarFromBuff(bytes.NewBuffer(trimmedPath), 108)
-		}
-		if err != nil {
-			return nil, fmt.Errorf("error parsing sockaddr_un: %v", err)
-		}
-		res["sun_path"] = sunPath
-	case 2: // AF_INET
-		/*
-			http://man7.org/linux/man-pages/man7/ip.7.html
-			struct sockaddr_in {
-				sa_family_t    sin_family; // address family: AF_INET
-				in_port_t      sin_port;   // port in network byte order
-				struct in_addr sin_addr;   // internet address
-				// byte        padding[8]; //https://elixir.bootlin.com/linux/v4.20.17/source/include/uapi/linux/in.h#L232
-			};
-			struct in_addr {
-				uint32_t       s_addr;     // address in network byte order
-			};
-		*/
-		var port uint16
-		err = binary.Read(buff, binary.BigEndian, &port)
-		if err != nil {
-			return nil, fmt.Errorf("error parsing sockaddr_in: %v", err)
-		}
-		res["sin_port"] = strconv.Itoa(int(port))
-		var addr uint32
-		err = binary.Read(buff, binary.BigEndian, &addr)
-		if err != nil {
-			return nil, fmt.Errorf("error parsing sockaddr_in: %v", err)
-		}
-		res["sin_addr"] = PrintUint32IP(addr)
-		_, err := readByteSliceFromBuff(buff, 8)
-		if err != nil {
-			return nil, fmt.Errorf("error parsing sockaddr_in: %v", err)
-		}
-	case 10: // AF_INET6
-		/*
-			struct sockaddr_in6 {
-				sa_family_t     sin6_family;   // AF_INET6
-				in_port_t       sin6_port;     // port number
-				uint32_t        sin6_flowinfo; // IPv6 flow information
-				struct in6_addr sin6_addr;     // IPv6 address
-				uint32_t        sin6_scope_id; // Scope ID (new in 2.4)
-			};
-
-			struct in6_addr {
-				unsigned char   s6_addr[16];   // IPv6 address
-			};
-		*/
-		var port uint16
-		err = binary.Read(buff, binary.BigEndian, &port)
-		if err != nil {
-			return nil, fmt.Errorf("error parsing sockaddr_in6: %v", err)
-		}
-		res["sin6_port"] = strconv.Itoa(int(port))
-
-		var flowinfo uint32
-		err = binary.Read(buff, binary.BigEndian, &flowinfo)
-		if err != nil {
-			return nil, fmt.Errorf("error parsing sockaddr_in6: %v", err)
-		}
-		res["sin6_flowinfo"] = strconv.Itoa(int(flowinfo))
-		addr, err := readByteSliceFromBuff(buff, 16)
-		if err != nil {
-			return nil, fmt.Errorf("error parsing sockaddr_in6: %v", err)
-		}
-		res["sin6_addr"] = Print16BytesSliceIP(addr)
-		var scopeid uint32
-		err = binary.Read(buff, binary.BigEndian, &scopeid)
-		if err != nil {
-			return nil, fmt.Errorf("error parsing sockaddr_in6: %v", err)
-		}
-		res["sin6_scopeid"] = strconv.Itoa(int(scopeid))
-	}
-	return res, nil
-}
-
-// alert struct encodes a security alert message with a timestamp
-// it is used to unmarshal binary data and therefore should match (bit by bit) to the `alert_t` struct in the ebpf code.
-type alert struct {
-	Ts      uint64
-	Msg     uint32
-	Payload uint8
-}
-
-func readArgFromBuff(dataBuff io.Reader) (argTag, interface{}, error) {
-	var err error
-	var res interface{}
-	var argTag argTag
-	var argType argType
-	err = binary.Read(dataBuff, binary.LittleEndian, &argType)
-	if err != nil {
-		return argTag, nil, fmt.Errorf("error reading arg type: %v", err)
-	}
-	err = binary.Read(dataBuff, binary.LittleEndian, &argTag)
-	if err != nil {
-		return argTag, nil, fmt.Errorf("error reading arg tag: %v", err)
-	}
-	switch argType {
-	case u16T:
-		var data uint16
-		err = binary.Read(dataBuff, binary.LittleEndian, &data)
-		res = data
-	case intT:
-		var data int32
-		err = binary.Read(dataBuff, binary.LittleEndian, &data)
-		res = data
-	case uintT, devT, modeT:
-		var data uint32
-		err = binary.Read(dataBuff, binary.LittleEndian, &data)
-		res = data
-	case longT:
-		var data int64
-		err = binary.Read(dataBuff, binary.LittleEndian, &data)
-		res = data
-	case ulongT, offT, sizeT:
-		var data uint64
-		err = binary.Read(dataBuff, binary.LittleEndian, &data)
-		res = data
-	case pointerT:
-		var data uint64
-		err = binary.Read(dataBuff, binary.LittleEndian, &data)
-		res = uintptr(data)
-	case sockAddrT:
-		res, err = readSockaddrFromBuff(dataBuff)
-	case alertT:
-		var data alert
-		err = binary.Read(dataBuff, binary.LittleEndian, &data)
-		res = data
-	case credT:
-		var data external.SlimCred
-		err = binary.Read(dataBuff, binary.LittleEndian, &data)
-		res = data
-	case strT:
-		res, err = readStringFromBuff(dataBuff)
-	case strArrT:
-		var ss []string
-		var arrLen uint8
-		err = binary.Read(dataBuff, binary.LittleEndian, &arrLen)
-		if err != nil {
-			return argTag, nil, fmt.Errorf("error reading string array number of elements: %v", err)
-		}
-		for i := 0; i < int(arrLen); i++ {
-			s, err := readStringFromBuff(dataBuff)
-			if err != nil {
-				return argTag, nil, fmt.Errorf("error reading string element: %v", err)
-			}
-			ss = append(ss, s)
-		}
-		res = ss
-	case bytesT:
-		var size uint32
-		err = binary.Read(dataBuff, binary.LittleEndian, &size)
-		if err != nil {
-			return argTag, nil, fmt.Errorf("error reading byte array size: %v", err)
-		}
-		if size > 4096 {
-			return argTag, nil, fmt.Errorf("byte array size too big: %d", size)
-		}
-		res, err = readByteSliceFromBuff(dataBuff, int(size))
-	default:
-		// if we don't recognize the arg type, we can't parse the rest of the buffer
-		return argTag, nil, fmt.Errorf("error unknown arg type %v", argType)
-	}
-	if err != nil {
-		return argTag, nil, err
-	}
-	return argTag, res, nil
 }
